MAKE FILLWNODELIST RETURN POINTER
This is so if it advances pointer to next slot the calling function
knows about this. It also now takes pointer one past as a parameter so
we know when it has reached the end. We can compare addresses of current
pointer and end to see if we have reached the end.